ios/Goobusters/Goobusters/python-app/requirements-ios.txt
# Python packages needed for iOS client
# These must be compiled for iOS ARM64 or be pure Python

# Core (likely have iOS wheels)
flask
python-dotenv
httpx

# Data handling
numpy
pillow

# These are heavyweight and may not have iOS wheels:
# pandas - optional, for data analysis
# scipy - optional, for advanced processing
# opencv-contrib-python - REQUIRED for image processing, needs iOS build
# torch/torchvision - likely too large, tracking should be server-side

# Note: The iOS client primarily acts as a thin client that:
# 1. Syncs data from MD.ai
# 2. Proxies requests to the server
# 3. Displays the WebView UI
# Heavy processing (tracking) happens on the server
